pyproject.toml
[project]
name = "bankservice"
version = "0.1.0"
description = ""
authors = [
    {name = "Никита Быстров",email = "hallnik96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django (==4.2.17)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"

[tool.black]
line-length = 120
skip-string-normalization = true
